requirements.txt
flask
pypdf
google-genai
gunicorn